Change the uuid.UUID field to string
The UUID field in the Lifespan and Group will now be of type string.
This will still be a unique identifier generated from
github.com/google/uuid.
